+++ s4e-web/src/app/modal/components/dummy-modal/dummy-modal.component.ts
@@ -1,4 +1,4 @@
-import {Component, Inject, OnInit, Optional} from '@angular/core';
+import {Component, Inject} from '@angular/core';
 import {ModalComponent} from '../../utils/modal/modal.component';
 import {ModalService} from '../../state/modal.service';
 import {MODAL_DEF} from '../../modal.providers';